fix(config): change conflict_cosine_threshold f32→f64 to fix schema serialization
serde_json promotes f32 to f64 before serializing, causing 0.85f32 to serialize
as "0.8500000238418579" on Linux CI. Changing to f64 produces the canonical "0.85"
representation via Ryu's shortest-decimal algorithm.

Cast to f32 at the two resolves_conflict call sites since cosine_sim operates on
f32 embeddings. Update config-schema.json type_name accordingly.

+++ crates/anno-rag/src/config.rs
@@ -596,7 +596,7 @@ pub struct AnnoRagConfig {
         since = "0.10"
     )]
     #[serde(default = "default_conflict_cosine_threshold")]
-    pub conflict_cosine_threshold: f32,
+    pub conflict_cosine_threshold: f64,
 
     /// Maximum hop count for `Pipeline::graph_recall`. Default 2.
     /// Caps the BFS depth over `entity_refs`; higher values risk
@@ -639,7 +639,7 @@ fn default_compaction_min_age_secs() -> u64 {
     3600
 }
 
-fn default_conflict_cosine_threshold() -> f32 {
+fn default_conflict_cosine_threshold() -> f64 {
     0.85
 }
