+++ crates/anno-rag/src/pipeline.rs
@@ -997,7 +997,7 @@ impl Pipeline {
                 .memory_candidates_for_conflict(&m.entity_refs, m.session_id.as_deref())
                 .await?;
             for prior in &candidates {
-                if crate::conflict::resolves_conflict(&m, prior, self.cfg.conflict_cosine_threshold)
+                if crate::conflict::resolves_conflict(&m, prior, self.cfg.conflict_cosine_threshold as f32)
                 {
                     self.store
                         .memory_update_valid_to(&prior.id, m.created_at)
@@ -1103,7 +1103,7 @@ impl Pipeline {
                     if crate::conflict::resolves_conflict(
                         &enriched,
                         prior,
-                        self.cfg.conflict_cosine_threshold,
+                        self.cfg.conflict_cosine_threshold as f32,
                     ) {
                         self.store
                             .memory_update_valid_to(&prior.id, enriched.created_at)
